requirements.txt
setuptools==40.6.3
nltk==3.4
requests==2.20.1
Pillow==5.3.0
PyPDF2==1.26.0
docopt==0.6.2
pyenchant==2.0.0
progressbar2==3.39.2
termcolor==1.1.0
